store/requirements.txt
asgiref==3.4.1
Django==3.1.5
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
djangorestframework==3.12.2
drf-yasg==1.20.0